requirements.txt
streamlit
tensorflow-cpu>=2.16.0
numpy>=1.26.0,<2.0.0
Pillow
streamlit-option-menu